Implemented Auto Focus Upon Invalid Email or Password Entry Using Forward Refs
Necessary to autofocus the cursor after the user attempts to login while either the email or password entered is invalid. This is done using the useRef and useImpoerativeHandle hooks and wrapping the Input component function in the React.forwardRef method. This technique (sharing a components function with it's parent component), however, is an extremely niche case and should be avoided at almost any cost.